コントローラーの修正

+++ app/controllers/items_controller.rb
@@ -27,8 +27,10 @@ class ItemsController < ApplicationController
 
   def destroy
     @item = Item.find(params[:id])
-    if @item.destroy
-      redirect_to root_path
+    if current_user.id == @item.user_id
+      if @item.destroy
+        redirect_to root_path
+      end
     end
   end
